Using wait for multiplexing over HTTP/1.x
When a large burst of DNS requests came in and the
only HTTPS/2 connection was too many time in idle state,
curl closes it and opens brand new connection for every
single DNS request.
Obviously later it will keep only 1 and close all the others.
So it is better to wait for multiplexing.
When testing this setting, the DNS query latency values
did not raise.